requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
selenium>=4.0.0
python-dotenv>=0.19.0
pydantic>=1.8.0
